+++ Neuro/Optimizers/Adam.cs
@@ -6,7 +6,7 @@ namespace Neuro.Optimizers
     // Implementation based on https://github.com/tensorflow/tensorflow/blob/master/tensorflow/python/training/adam.py
     public class Adam : OptimizerBase
     {
-        public Adam(double lr = 0.001)
+        public Adam(float lr = 0.001f)
         {
             LearningRate = lr;
         }
@@ -21,12 +21,12 @@ namespace Neuro.Optimizers
 
             ++T;
 
-            var tempLearningRate = LearningRate * Math.Sqrt(1.0 - Math.Pow(Beta2, T)) / (1.0 - Math.Pow(Beta1, T));
+            var tempLearningRate = LearningRate * (float)Math.Sqrt(1.0 - Math.Pow(Beta2, T)) / (1.0f - (float)Math.Pow(Beta1, T));
 
             M.Mul(Beta1).Add(gradient.Map(g => (1 - Beta1) * g), M);
             V.Mul(Beta2).Add(gradient.Map(g => (1 - Beta2) * g * g), V);
 
-            return M.Div(V.Map(x => Math.Sqrt(x) + Epsilon)).Mul(tempLearningRate);
+            return M.Div(V.Map(x => (float)Math.Sqrt(x) + Epsilon)).Mul(tempLearningRate);
         }
 
         public override OptimizerBase Clone()
@@ -40,13 +40,13 @@ namespace Neuro.Optimizers
             //return $"Adam(lr={LearningRate}, beta1={Beta1}, beta2={Beta2}, epsilon={Epsilon})";
         }
 
-        private readonly double LearningRate;
-        private readonly double Beta1 = 0.9;
-        private readonly double Beta2 = 0.999;
-        private readonly double Epsilon = 1e-8;
+        private readonly float LearningRate;
+        private readonly float Beta1 = 0.9f;
+        private readonly float Beta2 = 0.999f;
+        private readonly float Epsilon = 1e-8f;
 
         private Tensor M;
         private Tensor V;
-        private double T;
+        private float T;
     }
 }
